requirements.txt
httpx~=0.18
Pillow
youtube-dl